fixed 'Clear Basket' button.

+++ app/Http/Controllers/BasketController.php
@@ -10,6 +10,7 @@ use Musicshop\Http\Controllers\Controller;
 
 class BasketController extends Controller
 {
+
     public function basket()
     {
         if (Request::isMethod('post')) {
@@ -36,21 +37,19 @@ class BasketController extends Controller
             Cart::update($rowId[0], $item->qty - 1);
         }
 
-        //remove remove item from the Cart collection
+        //remove item from the Cart collection
         if(Request::get('album_id') && (Request::get('remove')) == 'all'){
             $rowId = Cart::search(array('id' => Request::get('album_id')));
             Cart::remove($rowId[0]);
+        }
+
+        //remove all item form the Cart collection
+        if(Request::get('basket') == 'clear'){
+            Cart::destroy();
         }
         
         $basket = Cart::content();
 
         return view('basket.index', compact('basket'));
-    }
-
-    public function clear()
-    {
-        Cart::destroy();
-
-        return back();
     }
 }
